Eliminate unnecessarily added util classes
