+++ ksml/src/main/java/io/axual/ksml/operation/parser/OperationParser.java
@@ -20,20 +20,22 @@ package io.axual.ksml.operation.parser;
  * =========================LICENSE_END==================================
  */
 
+import io.axual.ksml.data.schema.StructSchema;
 import io.axual.ksml.definition.StateStoreDefinition;
+import io.axual.ksml.definition.parser.StateStoreDefinitionParser;
 import io.axual.ksml.dsl.KSMLDSL;
+import io.axual.ksml.exception.ParseException;
 import io.axual.ksml.generator.TopologyResources;
 import io.axual.ksml.metric.MetricTags;
 import io.axual.ksml.operation.BaseOperation;
 import io.axual.ksml.operation.DualStoreOperationConfig;
 import io.axual.ksml.operation.OperationConfig;
 import io.axual.ksml.operation.StoreOperationConfig;
-import io.axual.ksml.parser.NamedObjectParser;
-import io.axual.ksml.parser.StructsParser;
-import io.axual.ksml.parser.TopologyResourceAwareParser;
-import io.axual.ksml.store.StoreParserUtil;
+import io.axual.ksml.parser.*;
 import io.axual.ksml.store.StoreType;
 import lombok.Getter;
+
+import java.util.List;
 
 @Getter
 public abstract class OperationParser<T extends BaseOperation> extends TopologyResourceAwareParser<T> implements NamedObjectParser {
@@ -62,13 +64,35 @@ public abstract class OperationParser<T extends BaseOperation> extends TopologyR
         return new StoreOperationConfig(name != null ? resources().getUniqueOperationName(name) : resources().getUniqueOperationName(tags), tags, store);
     }
 
-    protected StructsParser<StateStoreDefinition> storeField(boolean required, String doc, StoreType expectedStoreType) {
-        return StoreParserUtil.storeField(KSMLDSL.Operations.STORE_ATTRIBUTE, required, doc, expectedStoreType, resources());
-    }
-
     protected DualStoreOperationConfig dualStoreOperationConfig(String name, MetricTags tags, StateStoreDefinition store1, StateStoreDefinition store2) {
         name = validateName("Store", name, defaultShortName(), true);
         return new DualStoreOperationConfig(name != null ? resources().getUniqueOperationName(name) : resources().getUniqueOperationName(tags), tags, store1, store2);
+    }
+
+    protected StructsParser<StateStoreDefinition> storeField(boolean required, String doc, StoreType expectedStoreType) {
+        return storeField(KSMLDSL.Operations.STORE_ATTRIBUTE, required, doc, expectedStoreType);
+    }
+
+    protected StructsParser<StateStoreDefinition> storeField(String childName, boolean required, String doc, StoreType expectedStoreType) {
+        final var stateStoreParser = new StateStoreDefinitionParser(expectedStoreType, false);
+        final var resourceParser = new TopologyResourceParser<>("state store", childName, doc, (name, context) -> resources().stateStore(name), stateStoreParser);
+        final var schemas = required ? resourceParser.schemas() : optional(resourceParser).schemas();
+        return new StructsParser<>() {
+            @Override
+            public StateStoreDefinition parse(ParseNode node) {
+                stateStoreParser.defaultShortName(node.name());
+                stateStoreParser.defaultLongName(node.longName());
+                final var resource = resourceParser.parse(node);
+                if (resource != null && resource.definition() instanceof StateStoreDefinition def) return def;
+                if (!required) return null;
+                throw new ParseException(node, "Required state store is not defined");
+            }
+
+            @Override
+            public List<StructSchema> schemas() {
+                return schemas;
+            }
+        };
     }
 
     @Override
